+++ ceph_medic/tests/util/test_hosts.py
@@ -1,11 +1,15 @@
+import json
 import pytest
 from ceph_medic.util import hosts, configuration
 import ceph_medic
 from textwrap import dedent
 
 
-def failed_check():
-    raise RuntimeError('command failed')
+def failed_check(raise_=True):
+    if raise_:
+        raise RuntimeError('command failed')
+    else:
+        return dict(stdout='', stderr='', code=1)
 
 
 class TestContainerPlatform(object):
@@ -78,3 +82,46 @@ class TestContainerPlatform(object):
             'oc', '--context', '87', '--request-timeout=5', 'get', '-n',
             'rook-ceph', 'pods', '-o', 'json'
         ]
+
+
+class TestBasicContainers(object):
+    binaries = ['docker', 'podman']
+
+    @pytest.mark.parametrize('binary', binaries)
+    def test_executable_fails(
+            self, binary, monkeypatch, make_nodes, capsys):
+        monkeypatch.setattr(hosts.config, 'nodes', make_nodes(mgrs=['mgr0']))
+        monkeypatch.setattr(
+            hosts.ceph_medic.connection, 'get_connection',
+            lambda *a, **k: None)
+        monkeypatch.setattr(
+            hosts.process, 'check', lambda *a: failed_check(False))
+        hosts.basic_containers(binary)
+        stdout, stderr = capsys.readouterr()
+        assert 'Unable to list containers on host mgr0' in stdout
+
+    @pytest.mark.parametrize('binary', binaries)
+    def test_inspection(
+            self, binary, monkeypatch, make_nodes, stub_check, capsys):
+        monkeypatch.setattr(ceph_medic.config, 'cluster_name', 'ceph')
+        monkeypatch.setattr(hosts.config, 'nodes', make_nodes(mgrs=['mgr0']))
+        monkeypatch.setattr(
+            hosts.ceph_medic.connection, 'get_connection',
+            lambda *a, **k: None)
+        fake_list = json.dumps([{'Names': 'mgr0-container'}])
+        fake_mgr = json.dumps([{
+            'Name': 'mgr0-container',
+            'Config': {
+                'Env': [
+                    'CLUSTER=ceph',
+                    'CEPH_DAEMON=MGR',
+                ]
+            }
+        }])
+        stub_check([
+            ([fake_mgr], [''], 0),
+            ([fake_list], [''], 0),
+        ])
+        result = hosts.basic_containers(binary)
+        assert result['mgrs'][0]['host'] == 'mgr0'
+        assert result['mgrs'][0]['container'] == 'mgr0-container'