+++ ceph_medic/util/hosts.py
@@ -1,5 +1,6 @@
 import json
 import logging
+import ceph_medic.connection
 from ceph_medic import config, terminal
 from remoto import connection, process
 
@@ -87,4 +88,60 @@ def container_platform(platform='openshift'):
     for key, value in dict(base_inventory).items():
         if not value:
             base_inventory.pop(key)
+    return base_inventory
+
+
+def basic_containers(deployment_type):
+    base_inventory = {
+        'rgws': [], 'mgrs': [], 'mdss': [], 'clients': [], 'osds': [],
+        'mons': []
+    }
+    label_map = {
+        'OSD': 'osds',
+        'OSD_CEPH_VOLUME_ACTIVATE': 'osds',
+        'MON': 'mons',
+        'MGR': 'mgrs',
+        'MDS': 'mdss',
+        'RGW': 'rgws',
+    }
+    metal_hosts = set()
+    for nodes in config.nodes.values():
+        for node in nodes:
+            metal_hosts.add(node['host'])
+    for host in metal_hosts:
+        logger.debug("listing containers for host %s", host)
+        cmd = [deployment_type, 'container', 'ls', '--format', 'json',
+               '--no-trunc']
+        conn = ceph_medic.connection.get_connection(
+            host, deployment_type='ssh')
+        out, err, code = process.check(conn, cmd)
+        if code:
+            terminal.error("Unable to list containers on host %s" % host)
+            continue
+        container_list = json.loads(''.join(out))
+        if not container_list:
+            terminal.warning("Host %s had no containers" % host)
+            continue
+        for container in container_list:
+            cmd = [deployment_type, 'container', 'inspect', container['Names']]
+            out, err, code = process.check(conn, cmd)
+            if code:
+                terminal.error(
+                    "Unable to inspect container %s on host %s" %
+                    (container['Names'], host)
+                )
+                continue
+            detail = json.loads(''.join(out))[0]
+            env = dict(
+                [s.split('=', 1) for s in detail['Config']['Env']])
+            if 'CEPH_DAEMON' not in env:
+                continue
+            if env.get('CLUSTER') != config.cluster_name:
+                continue
+            role = env['CEPH_DAEMON']
+            if role not in label_map:
+                continue
+            base_inventory[label_map[role]].append(
+                {'host': host, 'container': detail['Name'], 'group': None}
+            )
     return base_inventory