+++ docs/source/changelog.rst
@@ -1,3 +1,9 @@
+1.0.6
+-----
+* Docker, podman container support
+* Fix broken SSH config option
+* Fix querying the Ceph version via admin socket on newer Ceph versions
+
 1.0.5
 -----
 27-Jun-2019